feat, refactor: Implementation of the 'delete' button.
feat --> Implementation of the 'delete' button, but also little changes
en the list of 'specialValues' to be process in the 'calculatorFunctionality.js' file.
In the CSS styling file (calculatorStyle.css) there was a little change,
making the 'del' button undraggable and unselectable, just clickable.

refactor --> Refactor in the 'bottomScreenPrint' function in the JS
script, about the 'comma' button.