feat, style: Now when the user over-pass the 21 numbers/symbols, the skull will extend the calculator to make it all readable. And little modifications in the CSS styling file.
feat --> Improvement in the 'skullPositionFunction' in the
calculatorFunctionality.js script.
When the calculator's screen has more than 21 characters/numbers/symbols
it will begin to be extended to the right side.

style --> Change of all 'left' and 'width' Measurement type properties from
percentage to pixels inside of the calculator's section. Keeping the same size.